合并分支

+++ src/views/shoppingCart/src/selfGoods.vue
@@ -13,11 +13,13 @@
                   w-checkbox(v-model="i.checked", @change="selectedChange(i.checked,storeItem.store_name,false,i.sc_id)")
                 .img
                   img(:src="i.logo | img-filter")
-                  p(v-if="i.goods_num > i.storage_num") 仅剩{{i.storage_num}}件
+                  p(v-if="i.storage_num === 0")
+                    span 库存
+                    span 不足
                 .info(@click.stop="")
                   .text
-                    .name {{i.gi_name}}
-                    .spec
+                    .name(@click="goGoodsDetail(i.gspu_id)") {{i.gi_name}}
+                    .spec(@click="openSpecChange(i.gspu_id, i)")
                       span(v-for="(item,index) in i.specVOList") {{item.gspec_value}} {{index < i.specVOList.length-1? ';':''}}
                       img(src="../../../assets/img/ic_page_xljt@2x.png")
                     w-counter.counter(v-model="i.goods_num", @click.stop="", @change="countChange(i.sc_id,i.gsku_id,i.goods_num)", :min="1", :max="i.storage_num", width="2rem", height="20px")
@@ -41,16 +43,20 @@
             .bottomOperation
               .more
                 p(:class="{opc0:i.difference_price<=0}") 比加入时降{{i.difference_price | price-filter}}
-                img(src="../../../assets/img/diandian.png")
-                .moreOperation
-                  .sanjiao
-                  ul.buttons
-                    li(@click="changeType(storeIndex,index,i)")
-                      img(src="../../../assets/img/shoppingCartChange.png")
-                      p 快递配送
-                    li(@click="deleteGoods(i.sc_id, storeIndex, index)")
-                      img(src="../../../assets/img/shoppingCartDelete.png")
-                      p 删除
+                .moreRight
+                  img(src="../../../assets/img/shoppingCartMore.png")
+                  .moreOperation
+                    .sanjiao
+                    ul.buttons
+                      li(@click="changeType(storeIndex,index,i)", v-if="i.storage_num>0")
+                        img(src="../../../assets/img/shoppingCartChange.png")
+                        p 快递配送
+                      li(@click="changeStore(storeIndex,index,i)", v-if="i.storage_num>0")
+                        img(src="../../../assets/img/shoppingCartChange.png")
+                        p 提货门店
+                      li(@click="deleteGoods(i.sc_id, storeIndex, index)")
+                        img(src="../../../assets/img/shoppingCartDelete.png")
+                        p 删除
         .bottom
           .left <img src="../../../assets/img/location.png"/>{{storeItem.store_address}}
           .right
@@ -64,7 +70,8 @@
         animateName: 'leftOut',
         watchFlag: true,
         flag: true,
-        isdefault: false
+        isdefault: false,
+        exitAll: true // 为true才可以由子选框反选全选框
       }
     },
     props: {
@@ -80,6 +87,9 @@
     },
     watch: {
       allClick(val) {
+        if (!this.$store.state.exitAllChecked) {
+          return
+        }
         let scId = []
         this.goodsList.forEach((now) => {
           now.checked = this.allClick
@@ -102,6 +112,10 @@
       }
     },
     methods: {
+      // 打开修改规格
+      openSpecChange (id,spec) {
+        this.$parent.openSpecChange(id, spec)
+      },
       // 前往商品详情
       goGoodsDetail(id) {
         this.$router.push({
@@ -120,6 +134,18 @@
           }
         }
         this.$emit('tab', data, fun)
+
+      },
+      changeStore(storeIndex, index, data) {
+        this.animateName = 'leftOut'
+        let fun = () => {
+          this.goodsList[storeIndex].shoppingCartVOList.splice(index, 1)
+          if (this.goodsList[storeIndex].shoppingCartVOList.length<1) {
+            this.goodsList.splice(storeIndex,1)
+          }
+        }
+        this.$store.commit('getSkuId',data.gsku_id)
+        this.$emit('change', data, fun)
 
       },
       edit(k, index) {
@@ -225,10 +251,10 @@
         })
         // 判断已选数据与总数据长度
         if (checked.length === allGoodsLen) {
-
           this.$store.commit('allCheckedChange', true)
         } else {
-          // this.$store.commit('allCheckedChange', false)
+          this.$store.commit('exitAllCheckedChange', false)
+          this.$store.commit('allCheckedChange', false)
         }
       },
       // 商品数量变化
@@ -249,24 +275,34 @@
         })
       },
       deleteGoods(id, storeIndex, index) {
-        this.animateName = 'fadeOut'
-        this.goodsList[storeIndex].shoppingCartVOList.splice(index, 1)
-        if (this.goodsList[storeIndex].shoppingCartVOList.length<1) {
-          this.goodsList.splice(storeIndex,1)
-        }
-        let self = this
-        self.$ajax({
-          method: 'delete',
-          url: self.$apiApp + 'shoppingCart/shoppingCart/delete',
-          params: {
-            scIdArray: id
+        this.$confirm({
+          title: '删除购物商品',
+          message: '确定要删除么',
+          confirm: () => {
+            this.animateName = 'fadeOut'
+            this.goodsList[storeIndex].shoppingCartVOList.splice(index, 1)
+            if (this.goodsList[storeIndex].shoppingCartVOList.length<1) {
+              this.goodsList.splice(storeIndex,1)
+            }
+            let self = this
+            self.$ajax({
+              method: 'delete',
+              url: self.$apiApp + 'shoppingCart/shoppingCart/delete',
+              params: {
+                scIdArray: id
+              },
+            }).then(function (response) {
+              self.$emit('clear')
+              // let goodsNum = self.$store.state.shoppingCartGoodsNum
+              // goodsNum.carryNum-=1
+              // self.$store.commit('shoppingCartGoodsNumChange',goodsNum)
+            })
           },
-        }).then(function (response) {
-          self.$emit('clear')
-          // let goodsNum = self.$store.state.shoppingCartGoodsNum
-          // goodsNum.carryNum-=1
-          // self.$store.commit('shoppingCartGoodsNumChange',goodsNum)
+          noConfirm: () => {
+
+          }
         })
+
       }
     }
   }
@@ -329,15 +365,19 @@
 
   .img p {
     padding-left: 2px;
-    font-size: .2rem;
+    font-size: .35rem;
+    letter-spacing: 4px;
     position: absolute;
     bottom: 0;
     left: 0;
-    height: .5rem;
-    line-height: .5rem;
+    height: 100%;
     width: 100%;
     background-color: rgba(0, 0, 0, 0.5);
     color: white;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
   }
 
   /*  */
@@ -498,7 +538,10 @@
     align-items: center;
     justify-content: space-between;
   }
-  .bottomOperation> .more:hover .moreOperation {
+  .moreRight {
+    position: relative;
+  }
+  .bottomOperation> .more>.moreRight:hover .moreOperation {
     display: block;
   }
   .bottomOperation> .more> p{
@@ -506,15 +549,16 @@
     border: solid 1px #F70057;
     padding: 1px .2rem;
   }
-  .bottomOperation> .more> img{
+  .bottomOperation> .more> .moreRight>img{
     height: .4rem;
     margin-right: .4rem;
   }
-  .bottomOperation> .more> .moreOperation {
+  .bottomOperation> .more .moreRight .moreOperation {
     display: none;
     position: absolute;
     right: 0;
-    top: .7rem;
+    top: .3rem;
+    z-index: 99;
   }
   .moreOperation>.buttons {
     width: 2.5rem;
@@ -529,7 +573,7 @@
     align-items: center;
     padding: 0 .2rem;
   }
-  .moreOperation>.buttons li img{
+  .moreOperation>.buttons>li img{
     height: .4rem;
   }
   .moreOperation>.buttons li p{
@@ -565,7 +609,7 @@
   .leftOut-enter, .leftOut-leave-to
     /* .slide-fade-leave-active for below version 2.1.8 */
   {
-    transform: translate(100%, -1000%) scale(.1, .1);
+    transform: translate(-100%, -1000%) scale(.1, .1);
     opacity: 0;
   }
